Refactor MaxPooling.
Pull out most of the logic from MaxPooling and put it in
Pooling so that I can reuse it for AveragePooling and SumPooling.

The new base Pooling class contains some properties that can
modify input_dim, in order to mimic the interface of Convolutional
and Convolutional{Activation,Layer} bricks. It also provides a
more complete access to the various arguments of max_pool_2d.