Fix Pooling output dims w/ unspecified input dims
The Pooling classes break when trying to provide output dimensions when
the input dimensions are unspecified. This is particularly important for
use with ConvolutionalSequence. This fixes that with a simple check on
the input dimensions rather than passing the Nones down to Theano's
Pool.out_shape(), which won't handle them. Also included is a regression
test, as well as an addition to the ConvolutionalSequence regression
test.

Fixes #1048.