Use tensor.nnet.relu for ReLU calculations.
We punt the decision of how to implement max(0, x) or max(alpha * x, x) most efficiently to Theano, via `tensor.nnet.relu`. This is what Lasagne and Keras do, which between them have > 10x as many GitHub stars, so I am willing to treat this solution as "battle-tested".

(Finally) closes #960.

N.B. This was done via a quick edit on the GitHub web interface so the branch is on `mila-udem`, I'll delete it when merged.